fix: tests

+++ packages/wabe/src/files/hookDeleteFile.ts
@@ -1,9 +1,6 @@
 import type { HookObject } from '../hooks/HookObject'
 
 const deleteFile = async (hookObject: HookObject<any, any>) => {
-  if (!hookObject.context.wabe.controllers.file)
-    throw new Error('No file adapter found')
-
   const schema = hookObject.context.wabe.config.schema?.classes?.find(
     (currentClass) => currentClass.name === hookObject.className,
   )
@@ -16,6 +13,9 @@ const deleteFile = async (hookObject: HookObject<any, any>) => {
       const fileName = hookObject.originalObject?.[fieldName]?.name as string
 
       if (!fileName) return
+
+      if (!hookObject.context.wabe.controllers.file)
+        throw new Error('No file adapter found')
 
       await hookObject.context.wabe.controllers.file?.deleteFile(fileName)
     })
